added LoginContent components that displays in Login view with Header
